modularity integration working
