--- RISC_V_modular/configs/example.json
@@ -1,31 +1,0 @@
-{
-  "STAGE_IF": ["HazardDetectionRV32I.core.IFBarrier.io_pc_out", "PipelinedRV32I.core.ifBarrier.io_pc_out"],
-  "STAGE_ID": ["HazardDetectionRV32I.core.IDBarrier.pcReg", "PipelinedRV32I.core.idBarrier.pcReg"],
-  "STAGE_EX": ["HazardDetectionRV32I.core.EXBarrier.pcReg", "PipelinedRV32I.core.exBarrier.pcReg"],
-  "STAGE_MEM": ["HazardDetectionRV32I.core.MEMBarrier.pcReg", "PipelinedRV32I.core.memBarrier.pcReg"],
-  "STAGE_WB": ["HazardDetectionRV32I.core.WBBarrier.pcReg", "PipelinedRV32I.core.wbBarrier.pcReg"],
-
-  "INSTR_IF": ["HazardDetectionRV32I.core.IFBarrier.io_inst_out", "PipelinedRV32I.core.ifBarrier.io_inst_out"],
-  "INSTR_ID": ["HazardDetectionRV32I.core.IDBarrier.instReg", "PipelinedRV32I.core.idBarrier.instReg"],
-  "INSTR_EX": ["HazardDetectionRV32I.core.EXBarrier.instReg", "PipelinedRV32I.core.exBarrier.instReg"],
-  "INSTR_MEM": ["HazardDetectionRV32I.core.MEMBarrier.instReg", "PipelinedRV32I.core.memBarrier.instReg"],
-  "INSTR_WB": ["HazardDetectionRV32I.core.WBBarrier.instReg", "PipelinedRV32I.core.wbBarrier.instReg"],
-
-  "EX_OPERAND_A": ["HazardDetectionRV32I.core.EX.io_operandA", "PipelinedRV32I.core.exStage.io_data1_in"],
-  "EX_OPERAND_B": ["HazardDetectionRV32I.core.EX.io_operandB", "PipelinedRV32I.core.exStage.operandB"],
-  "EX_ALU_RESULT": ["HazardDetectionRV32I.core.EX.io_aluResult", "PipelinedRV32I.core.exStage.aluResult"],
-  "EX_UOP": ["HazardDetectionRV32I.core.EX.io_uop", "PipelinedRV32I.core.exStage.io_upo_in"],
-
-  "WB_RD": ["HazardDetectionRV32I.core.WB.io_rd", "PipelinedRV32I.core.WB.io_rd"],
-  "WB_DATA": ["HazardDetectionRV32I.core.WB.io_check_res", "PipelinedRV32I.core.io_check_res"],
-
-  "HAZARD_RS1_ADDR": ["HazardDetectionRV32I.core.IDBarrier.io_outRS1", "PipelinedRV32I.core.idBarrier.rs1"],
-  "HAZARD_RS2_ADDR": ["HazardDetectionRV32I.core.IDBarrier.io_outRS2", "PipelinedRV32I.core.idBarrier.rs2"],
-  "HAZARD_OPCODE": ["HazardDetectionRV32I.core.ID.opcode", "PipelinedRV32I.core.idStage.opcode"],
-  "HAZARD_RD_MEM_ADDR": ["HazardDetectionRV32I.core.EXBarrier.io_outRD"],
-  "HAZARD_RD_WB_ADDR": ["HazardDetectionRV32I.core.MEMBarrier.io_outRD"],
-  "HAZARD_FORWARD_A": ["HazardDetectionRV32I.core.FU.io_forwardA"],
-  "HAZARD_FORWARD_B": ["HazardDetectionRV32I.core.FU.io_forwardB"],
-
-  "REG_TEMPLATE": ["HazardDetectionRV32I.core.regFile.regFile_{i}", "PipelinedRV32I.core.idStage.rf.regFile_{i}"]
-}
