--- RISC_V_modular/configs/multicycle.json
@@ -1,37 +1,0 @@
-{
-  "STAGE": ["MultiCycleRV32Icore.stage"],
-  "PC": ["MultiCycleRV32Icore.PC"],
-  "FETCHED_INST": ["MultiCycleRV32Icore.inst"],
-  "RD_ADDR": ["MultiCycleRV32Icore.rdReg"],
-  "ALU_RESULT": ["MultiCycleRV32Icore.aluResult"],
-
-  
-  "INSTRUCTION": ["MultiCycleRV32Icore.instReg"],
-  "OPCODE": ["MultiCycleRV32Icore.opcode"],
-  "DECODED_RD": ["MultiCycleRV32Icore.rd"],
-  "FUNCT3": ["MultiCycleRV32Icore.funct3"],
-  "FUNCT7": ["MultiCycleRV32Icore.funct7"],
-  "IMM_SEXT": ["MultiCycleRV32Icore.immI_sext"],
-  "DECODED_RS1": ["MultiCycleRV32Icore.rs1"],
-  "DECODED_RS2": ["MultiCycleRV32Icore.rs2"],
-  "DECODED_RS1_DATA": ["MultiCycleRV32Icore.rs1Data"],
-  "DECODED_RS2_DATA": ["MultiCycleRV32Icore.rs2Data"],
-  "OPERAND_A": ["MultiCycleRV32Icore.operandA"],
-  "OPERAND_B": ["MultiCycleRV32Icore.operandB"],
-  "WB_DATA": ["MultiCycleRV32Icore.io_check_res"],
-
-
-  "IS_ADD":  ["MultiCycleRV32Icore.isADD"],
-  "IS_ADDI": ["MultiCycleRV32Icore.isADDI"],
-  "IS_SUB":  ["MultiCycleRV32Icore.isSUB"],
-  "IS_SLL":  ["MultiCycleRV32Icore.isSLL"],
-  "IS_SLT":  ["MultiCycleRV32Icore.isSLT"],
-  "IS_SLTU": ["MultiCycleRV32Icore.isSLTU"],
-  "IS_XOR":  ["MultiCycleRV32Icore.isXOR"],
-  "IS_SRL":  ["MultiCycleRV32Icore.isSRL"],
-  "IS_SRA":  ["MultiCycleRV32Icore.isSRA"],
-  "IS_OR":   ["MultiCycleRV32Icore.isOR"],
-  "IS_AND":  ["MultiCycleRV32Icore.isAND"],
-
-  "REG_TEMPLATE": "MultiCycleRV32Icore.regFile_{i}"
-}
