+++ run.py
@@ -1,0 +1,5 @@
+# run.py
+from src.main import main
+
+if __name__ == "__main__":
+    main()
